failed to fix spotify login button callback.
can’t figure out why the button isn’t working. regardless, this service doesn’t provide toplist… we are going to have to use userplaylistinfo, along with the fact that we will have to have parse running the ruby script in the background. This looks like a dead end to be honest. Coacoalib spotify might be the only way.